Release 0.10.0

+++ src/System/IO/SystemFolder.cs
@@ -1,18 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 
 #pragma warning disable CS1998
 
-namespace OwlCore.Storage.SystemIO;
+namespace OwlCore.Storage.System.IO;
 
 /// <summary>
 /// An <see cref="IFolder"/> implementation that uses System.IO.
 /// </summary>
-public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<SystemFile>, IFastFileMove<SystemFile>, IFastGetItem, IFastGetItemRecursive, IFastGetFirstByName, IFastGetRoot
+public class SystemFolder : IModifiableFolder, IChildFolder, ICreateCopyOf, IMoveFrom, IGetItem, IGetItemRecursive, IGetFirstByName, IGetRoot
 {
     private DirectoryInfo? _info;
 
@@ -23,6 +24,11 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     public SystemFolder(string path)
         : this(new DirectoryInfo(path))
     {
+        foreach (var c in global::System.IO.Path.GetInvalidPathChars())
+        {
+            if (path.Contains(c))
+                throw new FormatException($"Provided path contains invalid character '{c}'.");
+        }
     }
 
     /// <summary>
@@ -34,13 +40,13 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
         _info = info;
 
         // For consistency, always remove the trailing directory separator.
-        Path = info.FullName.TrimEnd(System.IO.Path.PathSeparator, System.IO.Path.DirectorySeparatorChar, System.IO.Path.AltDirectorySeparatorChar);
+        Path = info.FullName.TrimEnd(global::System.IO.Path.PathSeparator, global::System.IO.Path.DirectorySeparatorChar, global::System.IO.Path.AltDirectorySeparatorChar);
 
         if (!Directory.Exists(Path))
-            throw new FileNotFoundException($"Directory not found at path {Path}");
+            throw new FileNotFoundException($"Directory not found at path '{Path}'.");
 
         Id = Path;
-        Name = System.IO.Path.GetFileName(Path) ?? throw new ArgumentException($"Could not determine directory name from path {Path}");
+        Name = global::System.IO.Path.GetFileName(Path) ?? throw new ArgumentException($"Could not determine directory name from path '{Path}'.");
     }
 
     /// <summary>
@@ -117,7 +123,7 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     public Task<IStorableChild> GetItemRecursiveAsync(string id, CancellationToken cancellationToken = default)
     {
         if (!id.Contains(Path))
-            throw new FileNotFoundException($"The provided ID does not belong to an item in this folder.");
+            throw new FileNotFoundException($"The provided Id does not belong to an item in this folder.");
 
         // Since the path is used as the id, we can provide a fast method of getting a single item, without iterating.
         if (IsFile(id))
@@ -126,24 +132,24 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
         if (IsFolder(id))
             return Task.FromResult<IStorableChild>(new SystemFolder(id));
 
-        throw new ArgumentException($"Could not determine if the provided path is a file or folder. Path: {id}");
+        throw new ArgumentException($"Could not determine if the provided path is a file or folder. Path '{id}'.");
     }
 
     /// <inheritdoc />
     public Task<IStorableChild> GetItemAsync(string id, CancellationToken cancellationToken = default)
     {
         if (!id.Contains(Path))
-            throw new FileNotFoundException($"The provided ID does not belong to an item in this folder.");
+            throw new FileNotFoundException($"The provided Id does not belong to an item in this folder.");
 
         // Since the path is used as the id, we can provide a fast method of getting a single item, without iterating.
         if (IsFile(id))
         {
             // Capture file name, combine with known path. Forces reading from current folder only.
-            var fileName = System.IO.Path.GetFileName(id) ?? throw new ArgumentException($"Could not determine file name from id: {id}");
-            var fullPath = System.IO.Path.Combine(Path, fileName);
+            var fileName = global::System.IO.Path.GetFileName(id) ?? throw new ArgumentException($"Could not determine file name from Id '{id}'.");
+            var fullPath = global::System.IO.Path.Combine(Path, fileName);
 
             if (!File.Exists(fullPath))
-                throw new FileNotFoundException($"The provided ID does not belong to an item in this folder.");
+                throw new FileNotFoundException($"The provided Id does not belong to an item in this folder.");
 
             return Task.FromResult<IStorableChild>(new SystemFile(fullPath));
         }
@@ -151,19 +157,19 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
         if (IsFolder(id))
         {
             // Ensure containing directory matches current folder.
-            if (System.IO.Path.GetDirectoryName(id) != Path || !Directory.Exists(id))
-                throw new FileNotFoundException($"The provided ID does not belong to an item in this folder.");
+            if (global::System.IO.Path.GetDirectoryName(id) != Path || !Directory.Exists(id))
+                throw new FileNotFoundException($"The provided Id does not belong to an item in this folder.");
 
             return Task.FromResult<IStorableChild>(new SystemFolder(id));
         }
 
-        throw new FileNotFoundException($"Could not determine if the provided path exists, or whether it's a file or folder. Path: {id}");
+        throw new FileNotFoundException($"Could not determine if the provided path exists, or whether it's a file or folder. Id '{id}'.");
     }
 
     /// <inheritdoc/>
     public async Task<IStorableChild> GetFirstByNameAsync(string name, CancellationToken cancellationToken = default)
     {
-        return await GetItemAsync(System.IO.Path.Combine(Path, name), cancellationToken);
+        return await GetItemAsync(global::System.IO.Path.Combine(Path, name), cancellationToken);
     }
 
     /// <inheritdoc />
@@ -176,7 +182,7 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     public Task DeleteAsync(IStorableChild item, CancellationToken cancellationToken = default)
     {
         // Ensure containing directory matches current folder.
-        if (GetParentPath(item.Id).TrimEnd(System.IO.Path.DirectorySeparatorChar) != Path)
+        if (GetParentPath(item.Id).TrimEnd(global::System.IO.Path.DirectorySeparatorChar) != Path)
             throw new FileNotFoundException($"The provided item does not exist in this folder.");
 
         if (IsFolder(item.Id))
@@ -189,12 +195,17 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     }
 
     /// <inheritdoc />
-    public async Task<IChildFile> CreateCopyOfAsync(SystemFile fileToCopy, bool overwrite = default, CancellationToken cancellationToken = default)
+    public async Task<IChildFile> CreateCopyOfAsync(IFile fileToCopy, bool overwrite, CancellationToken cancellationToken, CreateCopyOfDelegate fallback)
     {
-        var newPath = System.IO.Path.Combine(Path, fileToCopy.Name);
+        // Check if the file is a SystemFile. If not, use the fallback.
+        if (fileToCopy is not SystemFile systemFile)
+            return await fallback(this, fileToCopy, overwrite, cancellationToken);
 
-        // If the target and destination are the same, there's no need to copy.
-        if (fileToCopy.Path == newPath)
+        // Handle using System.IO
+        var newPath = global::System.IO.Path.Combine(Path, systemFile.Name);
+
+        // If the source and destination are the same, there's no need to copy.
+        if (systemFile.Path == newPath)
             return new SystemFile(newPath);
 
         if (File.Exists(newPath))
@@ -205,23 +216,27 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
             File.Delete(newPath);
         }
 
-        File.Copy(fileToCopy.Path, newPath, overwrite);
+        File.Copy(systemFile.Path, newPath, overwrite);
 
         return new SystemFile(newPath);
     }
 
     /// <inheritdoc />
-    public async Task<IChildFile> MoveFromAsync(SystemFile fileToMove, IModifiableFolder source, bool overwrite = default,
-        CancellationToken cancellationToken = default)
+    public async Task<IChildFile> MoveFromAsync(IChildFile fileToMove, IModifiableFolder source, bool overwrite, CancellationToken cancellationToken, MoveFromDelegate fallback)
     {
-        var newPath = System.IO.Path.Combine(Path, fileToMove.Name);
+        // Check if the file is a SystemFile. If not, use the fallback.
+        if (fileToMove is not SystemFile systemFile)
+            return await fallback(this, fileToMove, source, overwrite, cancellationToken);
+
+        // Handle using System.IO
+        var newPath = global::System.IO.Path.Combine(Path, systemFile.Name);
         if (File.Exists(newPath) && !overwrite)
             return new SystemFile(newPath);
 
         if (overwrite)
             File.Delete(newPath);
 
-        File.Move(fileToMove.Path, newPath);
+        File.Move(systemFile.Path, newPath);
 
         return new SystemFile(newPath);
     }
@@ -229,7 +244,7 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     /// <inheritdoc />
     public Task<IChildFolder> CreateFolderAsync(string name, bool overwrite = false, CancellationToken cancellationToken = default)
     {
-        var newPath = System.IO.Path.Combine(Path, name);
+        var newPath = global::System.IO.Path.Combine(Path, name);
 
         try
         {
@@ -248,7 +263,7 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     /// <inheritdoc />
     public Task<IChildFile> CreateFileAsync(string name, bool overwrite = false, CancellationToken cancellationToken = default)
     {
-        var newPath = System.IO.Path.Combine(Path, name);
+        var newPath = global::System.IO.Path.Combine(Path, name);
 
         if (overwrite || !File.Exists(newPath))
             File.Create(newPath).Dispose();
@@ -263,22 +278,22 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
     }
 
     /// <inheritdoc />
-    public Task<IFolder?> GetRootAsync()
+    public Task<IFolder?> GetRootAsync(CancellationToken cancellationToken = default)
     {
         return Task.FromResult<IFolder?>(new SystemFolder(Info.Root));
     }
 
-    private static bool IsFile(string path) => System.IO.Path.GetFileName(path) is { } str && str != string.Empty && File.Exists(path);
+    private static bool IsFile(string path) => global::System.IO.Path.GetFileName(path) is { } str && str != string.Empty && File.Exists(path);
     private static bool IsFolder(string path) => Directory.Exists(path);
 
     string GetParentPath(string relativePath)
     {
         // Path.GetDirectoryName() treats strings that end with a directory separator as a directory. If there's no trailing slash, it's treated as a file.
         // Run it twice for folders. The first time only shaves off the trailing directory separator.
-        var parentDirectoryName = relativePath.EndsWith(System.IO.Path.DirectorySeparatorChar.ToString()) ? System.IO.Path.GetDirectoryName(System.IO.Path.GetDirectoryName(relativePath)) : System.IO.Path.GetDirectoryName(relativePath);
+        var parentDirectoryName = relativePath.EndsWith(global::System.IO.Path.DirectorySeparatorChar.ToString()) ? global::System.IO.Path.GetDirectoryName(global::System.IO.Path.GetDirectoryName(relativePath)) : global::System.IO.Path.GetDirectoryName(relativePath);
 
         // It also doesn't return a string that has a path separator at the end.
-        return parentDirectoryName + System.IO.Path.DirectorySeparatorChar;
+        return parentDirectoryName + global::System.IO.Path.DirectorySeparatorChar;
     }
 
     string GetParentDirectoryName(string relativePath)
@@ -286,6 +301,6 @@ public class SystemFolder : IModifiableFolder, IChildFolder, IFastFileCopy<Syste
         var parentPath = GetParentPath(relativePath);
         var parentParentPath = GetParentPath(parentPath);
 
-        return parentPath.Replace(parentParentPath, "").TrimEnd(System.IO.Path.DirectorySeparatorChar);
+        return parentPath.Replace(parentParentPath, "").TrimEnd(global::System.IO.Path.DirectorySeparatorChar);
     }
 }
